refactor(dashboard-auth): move point-of-sale to own module
