+++ apps/dashboard/src/modules/point-of-sale/components/POSOverviewTable.vue
