+++ apps/dashboard/src/modules/point-of-sale/views/POSCreateView.vue
