+++ apps/dashboard/src/modules/point-of-sale/views/POSEditView.vue
