+++ apps/dashboard/src/modules/point-of-sale/views/POSInfoView.vue
